Shift serialization code to JSONSerializing.swift
Previously lived in JSONParsing.swift.
